[autosave] TestProject/BlockSheetTest.cs

+++ ServerUtility/TestProject/BlockSheetTest.cs
@@ -21,7 +21,7 @@ namespace TestProject
 {
     public static class ExtractExtension
     {
-        public static TCast ExtractDevice<TCast>(this IEnumerable<IDevice<IDeviceState<IPacketDeviceData>>> list, int parent, int module , int inter)
+        public static TCast ExtractDevice<TCast>(this IEnumerable<IDevice<IDeviceState<IPacketDeviceData>>> list, ushort parent, byte module , byte inter)
         {
             var id = new DeviceID(parent, module, inter);
             return (TCast)list.First(p => p.DeviceID == id);
